Safari-611.1.10/LayoutTests/platform/mac-wk1/webarchive/loading/test-loading-archive-subresource-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
frame "<!--frame1-->" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
frame "<!--frame1-->" - didFailProvisionalLoadWithError
main frame - didFinishLoadForFrame
 This test assumes the webarchive is not loaded.
